+++ apps/front/wraffle-webview/src/shared/ui/header/core/Title.tsx
@@ -1,0 +1,14 @@
+import type {TypographyProps} from '@wraffle/ui';
+import {Typography} from '@wraffle/ui';
+
+interface TitleProps extends TypographyProps {
+  children: string;
+}
+
+export const Title = ({children, size = 'h4', ...props}: TitleProps) => {
+  return (
+    <Typography size={size} {...props}>
+      {children}
+    </Typography>
+  );
+};
